DEV-1144 further split components for server side rendering, global ranking logic
